Module gossip_server::config

Structs

SeedNodeSettings

A seed node that can be used to bootstrap the gossip cluster.

ServerSettings

L0 Data Store settings.

UnboundedSenderEventHandler 🔒

Wraps an UnboundedSender to implement the notify::EventHandler trait. This allows sending file notification evnts to UnboundedSender instances.

Constants

CONFIG_ROOT 🔒

DYNAMIC_CONFIG_PATH 🔒

STATIC_CONFIG_PATH 🔒

Functions

dynamic_config_path 🔒

setup_settings_watch

Setup a watcher for the settings files and return a stream of settings changes.

static_config_path 🔒